src/requirements.txt
numpy==1.26.4
Pillow==10.3.0
streamlit==1.32.0
chromadb==0.5.0
